+++ src/main/kotlin/icu/windea/pls/config/config/delegated/CwtDeclarationConfig.kt
@@ -21,7 +21,7 @@ import icu.windea.pls.model.expressions.ParadoxDefinitionSubtypeExpression
  * 用于描述定义声明的结构，从而在定义声明中提供代码补全、代码检查等功能。
  *
  * 说明：
- * - 可在其中通过子类型表达式（[ParadoxDefinitionSubtypeExpression]）限定定义声明的结构。
+ * - 可在其中通过 `subtype[{expression}] = {...}` 指定需要匹配的子类型。其中 `{expression}` 为子类型表达式（[ParadoxDefinitionSubtypeExpression]）。支持嵌套使用。
  * - 可在其中引用别名规则（[CwtAliasConfig]）与单别名规则（[CwtSingleAliasConfig]），从而简化声明规则的编写。
  * - 切换类型（swapped type）的声明规则可以直接嵌套在对应的基础类型（base type）的声明规则中。
  *