+++ src/main/kotlin/icu/windea/pls/config/config/delegated/CwtTypeConfig.kt
@@ -133,13 +133,13 @@ private class CwtTypeConfigResolverImpl : CwtTypeConfig.Resolver, CwtConfigResol
 
     private fun doResolve(config: CwtPropertyConfig): CwtTypeConfig? {
         val name = config.key.removeSurroundingOrNull("type[", "]")?.orNull()?.optimized() ?: return null
-        val propElements = config.properties
-        if (propElements.isNullOrEmpty()) {
+        val propConfigs = config.properties
+        if (propConfigs.isNullOrEmpty()) {
             logger.warn("Skipped invalid type config (name: $name): Missing properties.".withLocationPrefix(config))
             return null
         }
 
-        val propGroup = propElements.groupBy { it.key }
+        val propGroup = propConfigs.groupBy { it.key }
         val paths = propGroup.getAll("path").mapNotNullTo(sortedSetOf()) { it.stringValue?.optimizedPath() }.optimized()
         val pathFile = propGroup.getOne("path_file")?.stringValue
         val pathExtension = propGroup.getOne("path_extension")?.stringValue?.optimizedPathExtension()
@@ -160,7 +160,7 @@ private class CwtTypeConfigResolverImpl : CwtTypeConfig.Resolver, CwtConfigResol
         val unique = propGroup.getOne("unique")?.booleanValue ?: false
         val severity = propGroup.getOne("severity")?.stringValue
         val graphRelatedTypes = config.optionData.graphRelatedTypes
-        val subtypes = propElements.mapNotNull { CwtSubtypeConfig.resolve(it) }.associateBy { it.name }.optimized()
+        val subtypes = propConfigs.mapNotNull { CwtSubtypeConfig.resolve(it) }.associateBy { it.name }.optimized()
         val localisation = propGroup.getOne("localisation")?.let { CwtTypeLocalisationConfig.resolve(it) }
         val images = propGroup.getOne("images")?.let { CwtTypeImagesConfig.resolve(it) }