+++ src/main/kotlin/icu/windea/pls/config/config/delegated/CwtTypeLocalisationConfig.kt
@@ -1,13 +1,13 @@
 package icu.windea.pls.config.config.delegated
 
+import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.util.UserDataHolderBase
 import icu.windea.pls.config.config.CwtConfigResolverScope
 import icu.windea.pls.config.config.CwtPropertyConfig
 import icu.windea.pls.config.configExpression.CwtLocalisationLocationExpression
 import icu.windea.pls.config.configExpression.CwtLocationExpression
-import icu.windea.pls.core.cache.CacheBuilder
+import icu.windea.pls.config.manipulation.CwtConfigCopyService
 import icu.windea.pls.core.optimized
-import icu.windea.pls.core.removeSurroundingOrNull
 import icu.windea.pls.model.expressions.ParadoxDefinitionSubtypeExpression
 
 /**
@@ -15,6 +15,9 @@ import icu.windea.pls.model.expressions.ParadoxDefinitionSubtypeExpression
  *
  * 用于定位对应类型的定义的相关本地化，以便在 UI 与各种提示信息中展示。
  * 具体而言，通过位置表达式（[CwtLocalisationLocationExpression]）进行定位，并最终解析为本地化。
+ *
+ * 说明：
+ * - 可在其中通过 `subtype[{expression}] = {...}` 指定需要匹配的子类型。其中 `{expression}` 为子类型表达式（[ParadoxDefinitionSubtypeExpression]）。支持嵌套使用。
  *
  * 路径定位：
  * - `types/type[{type}]/localisation`。其中 `{type}` 匹配定义类型。
@@ -40,7 +43,7 @@ import icu.windea.pls.model.expressions.ParadoxDefinitionSubtypeExpression
  * }
  * ```
  *
- * @property locationConfigs 子类型表达式与位置规则的配对列表。
+ * @property locationConfigGroup 子类型表达式与位置规则的配对列表。
  *
  * @see CwtLocationConfig
  * @see CwtLocationExpression
@@ -58,53 +61,44 @@ interface CwtTypeLocalisationConfig : CwtTypePresentationConfig {
 // region Implementations
 
 private class CwtTypeLocalisationConfigResolverImpl : CwtTypeLocalisationConfig.Resolver, CwtConfigResolverScope {
-    // no logger here (unnecessary)
+    private val logger = thisLogger()
 
     override fun resolve(config: CwtPropertyConfig): CwtTypeLocalisationConfig? = doResolve(config)
 
     private fun doResolve(config: CwtPropertyConfig): CwtTypeLocalisationConfig? {
-        val locationConfigs: MutableList<Pair<String?, CwtLocationConfig>> = mutableListOf()
-        val props1 = config.properties ?: return null
-        for (prop1 in props1) {
-            val subtypeName = prop1.key.removeSurroundingOrNull("subtype[", "]")?.optimized()
-            if (subtypeName != null) {
-                val props2 = prop1.properties ?: continue
-                for (prop2 in props2) {
-                    val locationConfig = CwtLocationConfig.resolve(prop2) ?: continue
-                    locationConfigs.add(subtypeName to locationConfig)
-                }
-            } else {
-                val locationConfig = CwtLocationConfig.resolve(prop1) ?: continue
-                locationConfigs.add(null to locationConfig)
-            }
-        }
-        return CwtTypeLocalisationConfigImpl(config, locationConfigs.optimized())
-    }
+        val locationConfigGroup = mutableMapOf<String, MutableList<CwtLocationConfig>>()
 
+        // #324
+        CwtConfigCopyService.flattenBySubtypeExpression(config) action@{ c, e ->
+            if (c !is CwtPropertyConfig) return@action
+            val locationConfig = CwtLocationConfig.resolve(c) ?: return@action
+            val key = e.optimized()
+            locationConfigGroup.getOrPut(key) { mutableListOf() }.add(locationConfig)
+        }
+
+        if (locationConfigGroup.isEmpty()) {
+            logger.warn("Skipped invalid type images config: Missing properties (after flatten).".withLocationPrefix(config))
+            return null
+        }
+        return CwtTypeLocalisationConfigImpl(config, locationConfigGroup.mapValues { (_, v) -> v.optimized() }.optimized())
+    }
 }
 
 private class CwtTypeLocalisationConfigImpl(
     override val config: CwtPropertyConfig,
-    override val locationConfigs: List<Pair<String?, CwtLocationConfig>> // (subtypeExpression, locationConfig)
+    override val locationConfigGroup: Map<String, List<CwtLocationConfig>>,
 ) : UserDataHolderBase(), CwtTypeLocalisationConfig {
-    private val configsCache = CacheBuilder().build<String, List<CwtLocationConfig>>()
-
-    override fun getConfigs(subtypes: List<String>): List<CwtLocationConfig> {
-        val cacheKey = subtypes.joinToString(",")
-        return configsCache.get(cacheKey) { computeConfigs(subtypes).optimized() }
-    }
-
-    private fun computeConfigs(subtypes: List<String>): MutableList<CwtLocationConfig> {
+    override fun getLocationConfigs(subtypes: List<String>): List<CwtLocationConfig> {
         val result = mutableListOf<CwtLocationConfig>()
-        for ((subtypeExpression, locationConfig) in locationConfigs) {
-            if (subtypeExpression == null || ParadoxDefinitionSubtypeExpression.resolve(subtypeExpression).matches(subtypes)) {
-                result.add(locationConfig)
+        for ((subtypeExpression, locationConfigs) in locationConfigGroup) {
+            if (subtypeExpression.isEmpty() || ParadoxDefinitionSubtypeExpression.resolve(subtypeExpression).matches(subtypes)) {
+                result.addAll(locationConfigs)
             }
         }
         return result
     }
 
-    override fun toString() = "CwtTypeLocalisationConfigImpl(locationConfigs=$locationConfigs)"
+    override fun toString() = "CwtTypeLocalisationConfigImpl(locationConfigs=$locationConfigGroup)"
 }
 
 // endregion