+++ src/main/kotlin/icu/windea/pls/config/config/delegated/CwtTypePresentationConfig.kt
@@ -10,14 +10,16 @@ import icu.windea.pls.cwt.psi.CwtProperty
  *
  * 用于定位对应类型的定义的相关本地化与图片，以便在 UI 与各种提示信息中展示。
  *
+ * @property locationConfigGroup 子类型表达式到一组位置规则的映射。如果子类型表达式为空，则表示适用于所有子类型。
+ *
  * @see CwtTypeLocalisationConfig
  * @see CwtTypeImagesConfig
  * @see CwtLocationConfig
  * @see CwtLocationExpression
  */
 sealed interface CwtTypePresentationConfig : CwtDelegatedConfig<CwtProperty, CwtPropertyConfig> {
-    val locationConfigs: List<Pair<String?, CwtLocationConfig>> // (subtypeExpression, locationConfig)
+    val locationConfigGroup: Map<String, List<CwtLocationConfig>>
 
     /** 按给定的 [subtypes] 合并与筛选后，返回生效的位置规则列表。 */
-    fun getConfigs(subtypes: List<String>): List<CwtLocationConfig>
+    fun getLocationConfigs(subtypes: List<String>): List<CwtLocationConfig>
 }