+++ CHANGELOG.md
@@ -22,7 +22,8 @@
 
 ## 2.1.9-dev
 
-- [x] 修复游戏类型发生变更时，分析数据可能未被正确刷新的问题 / Fixed an issue that analysis data may not be refreshed when game type is changed ([#326](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/32))
+- [x] 修复游戏类型发生变更时，分析数据可能未被正确刷新的问题 / Fixed an issue that analysis data may not be refreshed when game type is changed ([#326](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/326))
+- [x] 支持在类型展示规则中使用嵌套的 `subtype[x] = {...}` / Supported nested `subtype[x] = {...}` in type presentation config ([#324](https://github.com/DragonKnightOfBreeze/Paradox-Language-Support/issues/324))
 - [x] 其他优化与BUG修复 / Other optimizations and bug fixes
 
 ## 2.1.8 - 2026-05-10